+++ puzzle-game/game/systems/RenderSystem.cs
@@ -38,7 +38,7 @@ namespace puzzle_game.Game.Systems
             if (body.Shape is Rectangle)
             {
                 var rect = (Rectangle)body.Shape;
-                Raylib_cs.Raylib.DrawRectangle(body.X, body.Y, rect.Width, rect.Height, render.FillColor);
+                Raylib_cs.Raylib.DrawRectangle((int)body.X, (int)body.Y, rect.Width, rect.Height, render.FillColor);
             }
         }
     }
